engine/requirements.txt
groq
ddgs
requests
google-generativeai
colorama